+++ CHANGELOG.md
@@ -27,6 +27,7 @@ should change the heading of the (upcoming) version to include a major version b
 - Updated `resolveCondition` to default formData as empty object when evaluating if expression, fixing [#3706](https://github.com/rjsf-team/react-jsonschema-form/issues/3706)
 - Updated `retrieveSchemaInternal` to return failed merged allOf sub schemas for expandAllBranches flag, fixing [#3689](https://github.com/rjsf-team/react-jsonschema-form/issues/3700)
 - Updated `hashForSchema` to sort schema fields in consistent order before stringify to prevent different hash ids for the same schema
+- Updated `enumOptionsSelectValue` to allow picking falsy enumOptions, fixing [#3716](https://github.com/rjsf-team/react-jsonschema-form/issues/3716)
 
 ## @rjsf/bootstrap-4